[Modify] rename past files relevant with mvc patterns
