+++ src/legacy-mvc/controller/productController.js
