+++ src/legacy-mvc/index.js
@@ -1,9 +1,9 @@
 import { STORAGE } from '../constants/storage.js';
+import storage from '../utils/storage.js';
 import ChargeController from './controller/chargeController.js';
 import MenuController from './controller/menuController.js';
 import ProductController from './controller/productController.js';
 import VendingMachineModel from './model/vendingMachineModel.js';
-import storage from './utils/storage.js';
 import ChargeView from './view/chargeView.js';
 import MenuView from './view/menuView.js';
 import ProductView from './view/productView.js';
