+++ src/legacy-mvc/view/chargeView.js
