+++ src/legacy-mvc/view/menuView.js
