+++ src/legacy-mvc/view/productView.js
